requirements.txt
depthai>=2.18.0.0
matplotlib>=3.6.2
ngraph>=0.2.0
numpy>=1.22
pytest>=7.2.0
seaborn>=0.12.1
tqdm>=4.64.1
lakefs_client>=0.90.0
webdataset>=0.2.31
b2>=3.6.0
opencv-python>=4.7.0.68
PyYAML>=6.0
pandasql>=0.7.3
pycocotools>=2.0.6
boto3>=1.17.39
sqlalchemy==1.4.46
idx2numpy==1.2.3
fiftyone==0.21.0
label-studio-sdk>=0.0.28
google-cloud-storage>=2.10.0